backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhishek Chhabra <abhishekchhabra107@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
django-ninja = "^0.22.2"
django-enumfield = "^3.1"
furl = "^2.1.3"
html2text = "^2020.1.16"
ninja-schema = "^0.13.2"
django-extensions = "^3.2.3"
django-simple-history = "^3.3.0"
django-cors-headers = "^4.2.0"
python-dotenv = "^1.0.0"
gitpython = "^3.1.35"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
